refactor: Feature Based Patternにディレクトリ構成を変えるためにBoard,GameStatusを分離
