+++ src/App.tsx
@@ -1,86 +1,9 @@
 import { useEffect, useState } from "react";
-import {
-  BoardProps,
-  GameStatusProps,
-  PlayHistoryProps,
-  ToggleOrderProps,
-} from "./type";
+import { PlayHistoryProps, ToggleOrderProps } from "./type";
 
-import { Square } from "./Game/Board/Square/components/App";
+import { Board } from "./Game/Board/components/App";
 
-function Board({ xIsNext, board, onPlay }: BoardProps): JSX.Element {
-  const existWinner: number[][] | null = calculateWinner(board);
-
-  function handleClick(i: number, j: number): void {
-    if (board[i][j] || existWinner) {
-      return;
-    }
-    const newBoard = board.map((row) => row.slice());
-    if (xIsNext) {
-      newBoard[i][j] = "X";
-    } else {
-      newBoard[i][j] = "O";
-    }
-    onPlay(newBoard);
-  }
-
-  const boardRows = Array.from({ length: 3 });
-
-  return (
-    <>
-      {boardRows.map((_, i) => (
-        <div key={i} className="board-row">
-          {board[i].map((square, j) => {
-            const keyIndex = j + i * 3;
-            const isWinnersSquare =
-              existWinner &&
-              existWinner.some((elem) => elem[0] === i && elem[1] === j);
-            const squareStyle = isWinnersSquare
-              ? { backgroundColor: "pink" }
-              : {};
-
-            return (
-              <Square
-                key={keyIndex}
-                value={square}
-                onSquareClick={() => handleClick(i, j)}
-                style={squareStyle}
-              />
-            );
-          })}
-        </div>
-      ))}
-    </>
-  );
-}
-
-function GameStatus({ board, xIsNext }: GameStatusProps): JSX.Element {
-  const existWinner = calculateWinner(board);
-  const [gameStatus, setGameStatus] = useState<string>("");
-
-  useEffect(() => {
-    const newGameStatus = () => {
-      if (existWinner) {
-        return "Winner: " + board[existWinner[0][0]][existWinner[0][1]];
-      } else {
-        const isNullSquareExist = checkNullSquareExist(board);
-        if (isNullSquareExist) {
-          return "Next player: " + (xIsNext ? "X" : "O");
-        } else {
-          return "No Winner, No Loser";
-        }
-      }
-    };
-
-    setGameStatus(newGameStatus);
-  }, [board, xIsNext]);
-
-  return (
-    <>
-      <div className="gameStatus">{gameStatus}</div>
-    </>
-  );
-}
+import { GameStatus } from "./Game/GameStatus/components/App";
 
 function PlayHistory({ history, jumpTo }: PlayHistoryProps): JSX.Element {
   const [playHistory, setPlayHistory] = useState<JSX.Element[]>([]);
@@ -186,73 +109,4 @@ export default function Game(): JSX.Element {
       </div>
     </div>
   );
-}
-
-function calculateWinner(board: string[][]): number[][] | null {
-  const lines = [
-    [
-      [0, 0],
-      [0, 1],
-      [0, 2],
-    ],
-    [
-      [1, 0],
-      [1, 1],
-      [1, 2],
-    ],
-    [
-      [2, 0],
-      [2, 1],
-      [2, 2],
-    ],
-    [
-      [0, 0],
-      [1, 0],
-      [2, 0],
-    ],
-    [
-      [0, 1],
-      [1, 1],
-      [2, 1],
-    ],
-    [
-      [0, 2],
-      [1, 2],
-      [2, 2],
-    ],
-    [
-      [0, 0],
-      [1, 1],
-      [2, 2],
-    ],
-    [
-      [0, 2],
-      [1, 1],
-      [2, 0],
-    ],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [[a, x], [b, y], [c, z]] = lines[i];
-    if (
-      board[a][x] &&
-      board[a][x] === board[b][y] &&
-      board[a][x] === board[c][z]
-    ) {
-      return lines[i];
-    }
-  }
-  return null;
-}
-
-function checkNullSquareExist(Board: string[][]): boolean {
-  let isNullSquareExist = false;
-  Board.forEach((nthRow) => {
-    nthRow.forEach((nthSquare) => {
-      if (!nthSquare) {
-        isNullSquareExist = true;
-      }
-    });
-  });
-
-  return isNullSquareExist;
 }
